+++ m3u8/writer.go
@@ -749,7 +749,7 @@ func (p *MediaPlaylist) Encode() *bytes.Buffer {
 	}
 	if p.PartTargetDuration > 0 {
 		p.buf.WriteString("#EXT-X-PART-INF:PART-TARGET=")
-		p.buf.WriteString(strconv.FormatFloat(float64(p.PartTargetDuration), 'f', 3, 64))
+		p.buf.WriteString(strconv.FormatFloat(p.PartTargetDuration, 'f', 3, 64))
 		p.buf.WriteRune('\n')
 	}
 	p.buf.WriteString("#EXT-X-MEDIA-SEQUENCE:")
